backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
requests
email-validator
httpx